refactor: POST / GET 분리, 서비스가 Response를 반환하도록 변경
